ontology_system_v12/requirements.txt
pydantic>=2.0.0
pyyaml>=6.0
pytest>=7.0.0
yfinance>=0.2.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
